+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change default quadrature rule in `ParamsBiotSavart` from `GaussLegendre(2)`
   to `GaussLegendre(3)`.
 
+### Added
+
+- Add `longrange_truncate_spherical` option to `ParamsBiotSavart`. This should
+  only be used for testing.
+
 ## 0.17.1 - 2024-02-23
 
 ### Added